+++ OpenGL/src/IndexBuffer.h
@@ -7,7 +7,7 @@ private:
 	unsigned int m_count;
 
 public:
-	IndexBuffer(const unsigned int* data, size_t count);
+	IndexBuffer(const unsigned int* data, unsigned int count);
 	~IndexBuffer();
 
 	void Bind()const;
